+++ core/src/main/kotlin/ru/hse/se/team9/entities/GameObject.kt
@@ -1,0 +1,3 @@
+package ru.hse.se.team9.entities
+
+interface GameObject
