+++ core/src/main/kotlin/ru/hse/se/team9/entities/MapObject.kt
@@ -1,4 +1,4 @@
-package ru.hse.se.team9.gameobjects
+package ru.hse.se.team9.entities
 
 sealed class MapObject: GameObject
