--- core/src/main/kotlin/ru/hse/se/team9/gameobjects/GameObject.kt
@@ -1,3 +1,0 @@
-package ru.hse.se.team9.gameobjects
-
-interface GameObject
