+++ core/src/main/kotlin/ru/hse/se/team9/positions/Position.kt
@@ -1,3 +1,3 @@
-package ru.hse.se.team9.game.entities.map
+package ru.hse.se.team9.positions
 
 data class Position(val x: Int, val y: Int)
