+++ model/src/main/kotlin/ru/hse/se/team9/game/entities/hero/Hero.kt
@@ -1,5 +1,5 @@
 package ru.hse.se.team9.game.entities.hero
 
-import ru.hse.se.team9.gameobjects.GameObject
+import ru.hse.se.team9.entities.GameObject
 
 data class Hero(val stats: HeroStats): GameObject
