+++ model/src/main/kotlin/ru/hse/se/team9/game/entities/map/objects/HeroOnMap.kt
@@ -1,6 +1,6 @@
 package ru.hse.se.team9.game.entities.map.objects
 
 import ru.hse.se.team9.game.entities.hero.Hero
-import ru.hse.se.team9.game.entities.map.Position
+import ru.hse.se.team9.positions.Position
 
 class HeroOnMap(val hero: Hero, var position: Position)
